Rename package from app to gtasks_md
The distribution installed a top-level package literally named 'app',
which pollutes site-packages and collides with any other project that
made the same choice. Rename the import package to gtasks_md to match
the distribution name. All intra-package imports were already relative,
so only the tests and pyproject.toml needed updating.
