+++ tests/test_parser.py
@@ -1,8 +1,8 @@
 import unittest
 from inspect import cleandoc
 
-from app.parser import markdown_to_task_lists, task_lists_to_markdown
-from app.tasks import Task, TaskList, TaskStatus
+from gtasks_md.parser import markdown_to_task_lists, task_lists_to_markdown
+from gtasks_md.tasks import Task, TaskList, TaskStatus
 
 
 class TestMarkdownConversion(unittest.TestCase):
